requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.3.2
gunicorn==21.2.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
PyJWT==2.10.1
Werkzeug==2.3.7
zipp==3.23.0
